+++ starmath/source/ooxmlexport.cxx
@@ -496,7 +496,7 @@ void SmOoxmlExport::HandleBrace( const SmBraceNode* pNode, int nLevel )
             FSNS( XML_m, XML_val ), mathSymbolToString( pNode->ClosingBrace()).getStr(), FSEND );
 
     m_pSerializer->endElementNS( XML_m, XML_dPr );
-    for( unsigned int i = 0; i < subnodes.size(); ++i )
+    for( size_t i = 0; i < subnodes.size(); ++i )
     {
         m_pSerializer->startElementNS( XML_m, XML_e, FSEND );
         HandleNode( subnodes[ i ], nLevel + 1 );
